Refactor abstract payment service
