+++ processor/src/services/types/mock-payment.type.ts
@@ -1,15 +1,20 @@
-import { PaymentOutcome, PaymentRequestSchemaDTO } from '../../dtos/mock-payment.dto';
+import { PaymentRequestSchemaDTO } from '../../dtos/mock-payment.dto';
+import {CommercetoolsCartService, CommercetoolsPaymentService} from "@commercetools/connect-payments-sdk";
+
+export type MockPaymentServiceOptions = {
+  ctCartService: CommercetoolsCartService;
+  ctPaymentService: CommercetoolsPaymentService;
+};
 
 export type CreatePayment = {
   data: PaymentRequestSchemaDTO;
 };
 
-export type CreatePaymentRequest = {
-  data: PaymentRequestSchemaDTO;
-};
-
-export type MockPaymentProviderResponse = {
-  resultCode: PaymentOutcome;
-  pspReference: string;
-  paymentMethodType: string;
+export type MockPaymentRefundRequest = {
+  merchantAccount: string,
+  reference: request.payment.id,
+  amount: {
+    currency: request.amount.currencyCode,
+    value: request.amount.centAmount,
+  },
 };
